Rework CustomKnight protocol
Instead of sending the textures every time, the client now only sends
the hashes of the textures. The server then sends the list of hashes of
the missing textures. The client then finally sends the textures.

When a player is spawned, its hashes are sent with the player data.
Clients which don't have the textures in cache can ask the server for
them, and the server will fetch them from its own cache.

Having a local texture cache both on the client and on the server
minimizes bandwidth usage over time, because you only need to upload or
download a custom skin once.